Load existing improvement plan
